🚨 Extract `files` from `assets.root_paths`
TG-2031
