+++ src/youwol/backends/assets/routers/__init__.py
@@ -1,0 +1,2 @@
+# relative
+from .files import *
